Broke out types to its own file
